pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-analyzer"
dynamic = ["version"]
description = "Crypto quantitative research platform"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "scipy>=1.10.0",
    "PyYAML",
    "plotly",
    "streamlit",
    "requests",
    "matplotlib",
    "pyarrow",
]

[project.optional-dependencies]
api = ["fastapi>=0.100", "uvicorn[standard]>=0.20"]
postgres = ["sqlalchemy>=2.0", "psycopg2-binary>=2.9"]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.scripts]
crypto-analyzer = "crypto_analyzer.__main__:main"

[tool.setuptools.dynamic]
version = {attr = "crypto_analyzer.__version__"}

[tool.setuptools.packages.find]
include = ["crypto_analyzer*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"cli/*.py" = ["E402"]
"tools/*.py" = ["E402"]
"tests/*.py" = ["E402"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
